Extract map utils to separate file
